Refactor reading, encrypting, and writing tfvars
The functions that read, encrypt, and write the terraform.tfvars file
have been refactored by removing any hard-coded strings. The script
encrypts all secrets below the line "# Secrets below can be encrypted
using GCP-KMS, refer to the README for more information." and writes
to a new tfvars file. The original file is saved with a .backup postfix.
