+++ deployments/gcp/multi-region/terraform.tfvars.sample
@@ -1,7 +1,4 @@
 # Commented out lines represents defaults that can be changed
-####################
-#     General      #
-####################
 
 gcp_credentials_file = "/path/to/cred.json"
 gcp_project_id       = "your-project-1234"
@@ -21,17 +18,12 @@ gcp_service_account  = "service_account_name@<project_id>.iam.gserviceaccount.co
 # Optional: additional AD users to create
 # domain_users_list = "/path/to/domain_users_list.csv"
 
-
-####################
-#        CAC       #
-####################
 # Specify the regions, zones and number of connectors to create per region.
 # For example:
 #   cac_region_list         = ["us-west2",    "europe-west4",   "asia-southeast1"]
 #   cac_zone_list           = ["us-west2-b",  "europe-west4-b", "asia-southeast1-b"]
 #   cac_subnet_cidr_list    = ["10.0.1.0/24", "10.1.1.0/24",    "10.2.1.0/24"]
 #   cac_instance_count_list = [2,             3,                4]
-
 cac_region_list         = []
 cac_zone_list           = []
 cac_subnet_cidr_list    = []
@@ -43,12 +35,6 @@ cac_instance_count_list = []
 cac_admin_ssh_pub_key_file = "~/.ssh/id_rsa.pub"
 ssl_key  = "/path/to/privkey.pem"
 ssl_cert = "/path/to/fullchain.pem"
-
-
-####################
-#   Workstations   #
-####################
-# Specify the number of workstations created for each type.
 
 win_gfx_instance_count = 0
 # win_gfx_machine_type = "n1-standard-4"
@@ -76,30 +62,19 @@ centos_std_instance_count = 0
 
 centos_admin_ssh_pub_key_file = "~/.ssh/id_rsa.pub"
 
-
-####################
-#      GCP KMS     #
-####################
+###############
+#   Secrets   #
+###############
 # The secrets below may be supplied in 2 formats:
 #   1. plain text
 #   2. KMS encrypted, base64 encoded
-# For option 1, leave both keyring name and cryptokey name commented out.
-# For option 2, run the script located under tools using command "python3
-# gcp_kms_encrypto_secrets" which will overwrite this tfvars file with
-# your encrypted secrets and create an encrypted JSON for your CAM
-# credentials file. To set a custom keyring name and cryptokey name,
-# uncomment the two variables below and enter the new names, otherwise
-# a default name will be used by the script. The GCP service account 
-# specified above in "gcp_service_account" must be in the same project 
-# and have KMS decryptor permissions for this key.
+# For option 1, leave "kms_cryptokey_id" commented out.
+# For option 2, set "kms_cryptokey_id" to the KMS key used to encrypt the
+# secrets. The GCP service account specified above in "gcp_service_account"
+# must be in the same project and have KMS decryptor permissions for this key.
 
-# kms_keyring_name   = "terraform-keyring"
-# kms_cryptokey_name = "terraform-cryptokey"
+kms_cryptokey_id = "projects/<project-id>/locations/<location>/keyRings/<keyring-name>/cryptoKeys/<key-name>"
 
-
-####################
-#      Secrets     #
-####################
 # Note Windows password complexity requirements:
 # 1. Must not contain user's account name or display name
 # 2. Must have 3 of the following categories:
@@ -110,6 +85,7 @@ centos_admin_ssh_pub_key_file = "~/.ssh/id_rsa.pub"
 #    3. unicode characters
 # See: https://docs.microsoft.com/en-us/windows/security/threat-protection/security-policy-settings/password-must-meet-complexity-requirements
 
+# Secrets below can be encrypted using GCP-KMS, refer to the README for more information.
 dc_admin_password           = "SecuRe_pwd1"
 safe_mode_admin_password    = "SecuRe_pwd2"
 ad_service_account_password = "SecuRe_pwd3"
